[MOV] fingerprint: move to plugin
Fingerprint will be used by the evaluation in the next commit.
This moves the fingerprint computation to a plugin to make it
available to other plugins

Task: 6096737
